company changes

+++ app/Http/Controllers/EmployeeTypeController.php
@@ -5,8 +5,6 @@ namespace App\Http\Controllers;
 use App\Models\EmployeeType;
 use App\Http\Rules\EmployeeTypeRequest;
 use Illuminate\Http\JsonResponse;
-use Illuminate\Support\Facades\App;
-
 class EmployeeTypeController extends Controller
 {
     /**
@@ -15,7 +13,7 @@ class EmployeeTypeController extends Controller
     public function index()
     {
         $data = EmployeeType::all();
-        return api_response(200, 'Employee types received successfully', $data);
+        return api_response(true, 'Employee types received successfully', $data, 200);
     }
 
     /**
@@ -25,9 +23,9 @@ class EmployeeTypeController extends Controller
     {
         try {
             $employee_type = EmployeeType::create($request->validated());
-            return api_response(201, 'Employee type created successfully', $employee_type);
+            return api_response(true, 'Employee type created successfully', $employee_type, 201);
         } catch (Exception $e) {
-            return api_response(400, 'Internal server error', $e->getMessage(), JsonResponse::HTTP_INTERNAL_SERVER_ERROR);
+            return api_response(true, 'Internal server error', $e->getMessage(), JsonResponse::HTTP_INTERNAL_SERVER_ERROR);
         }
     }
 
@@ -36,7 +34,7 @@ class EmployeeTypeController extends Controller
      */
     public function show(EmployeeType $employee_type)
     {
-        return api_response(200, 'Employee type received successfully', $employee_type);
+        return api_response(true, 'Employee type received successfully', $employee_type, 200);
     }
 
     /**
@@ -45,10 +43,13 @@ class EmployeeTypeController extends Controller
     public function update(EmployeeTypeRequest $request, EmployeeType $employee_type)
     {
         try {
+            if (!$employee_type) {
+                return api_response(false, 'Employee type not found', $employee_type, 404);
+            }
             $employee_type->update($request->all());
-            return api_response(202, 'Employee type updated successfully', $employee_type);
+            return api_response(true, 'Employee type updated successfully', $employee_type, 202);
         } catch (Exception $e) {
-            return api_response(400, 'Internal server error', $e->getMessage(), JsonResponse::HTTP_INTERNAL_SERVER_ERROR);
+            return api_response(false, 'Internal server error', $e->getMessage(), JsonResponse::HTTP_INTERNAL_SERVER_ERROR);
         }
     }
 
@@ -57,7 +58,10 @@ class EmployeeTypeController extends Controller
      */
     public function destroy(EmployeeType $employee_type)
     {
+        if (!$employee_type) {
+            return api_response(false, 'Employee type not found', '', 404);
+        }
         $employee_type->delete();
-        return api_response(204, 'Employee type deleted');
+        return api_response(true, 'Employee type deleted', '', 200);
     }
 }
